requirements.txt
dataclasses==0.8
pyyaml==5.3.1
click==7.1.2
click-plugins==1.1.1
eli5==0.10.1
Faker==8.1.2
joblib==1.0.1 
jsonschema==3.2.0
marshmallow==3.11.1
marshmallow-dataclass==8.4.1
numpy==1.18.1
openpyxl==3.0.3
pytest==6.2.3
pytest-sugar==0.9.4
scikit-learn==0.24.1
xarray==0.16.1
xlrd==1.2.0
zict==2.0.0
matplotlib==3.2.1
